requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-groq
python-multipart
pydantic[email]
passlib 
bcrypt 
pyjwt 
aiosmtplib
passlib
annotated_doc
anyio
certifi